related: move `/service/ou/<uuid>/map` to separate blueprint, update docs
