+++ docs/api/detail_reading.rst
@@ -1,0 +1,18 @@
+.. automodule:: mora.service.detail_reading
+
+
+.. qrefflask:: mora.app:app
+   :blueprints: detail_reading
+   :order: path
+
+.. autoflask:: mora.app:app
+   :include-empty-docstring:
+   :order: path
+   :blueprints: detail_reading
+
+.. Indices and tables
+   ==================
+
+   * :ref:`genindex`
+   * :ref:`modindex`
+   * :ref:`search`
