+++ docs/api/detail_writing.rst
@@ -1,0 +1,18 @@
+.. automodule:: mora.service.detail_writing
+
+
+.. qrefflask:: mora.app:app
+   :blueprints: detail_writing
+   :order: path
+
+.. autoflask:: mora.app:app
+   :include-empty-docstring:
+   :order: path
+   :blueprints: detail_writing
+
+.. Indices and tables
+   ==================
+
+   * :ref:`genindex`
+   * :ref:`modindex`
+   * :ref:`search`
