+++ docs/api/related_unit.rst
@@ -1,0 +1,18 @@
+.. automodule:: mora.service.related
+
+
+.. qrefflask:: mora.app:app
+   :blueprints: related_unit
+   :order: path
+
+.. autoflask:: mora.app:app
+   :include-empty-docstring:
+   :order: path
+   :blueprints: related_unit
+
+.. Indices and tables
+   ==================
+
+   * :ref:`genindex`
+   * :ref:`modindex`
+   * :ref:`search`
